feat: add undefined behaviour sanitizer support

+++ src/kernel/main.c
@@ -194,7 +194,7 @@ void _start()
     init_pat();
 
     struct limine_file* initrd = NULL;
-    for (int i = 0; i < module_request.response->module_count; i++)
+    for (uint64_t i = 0; i < module_request.response->module_count; i++)
     {
         struct limine_file* file = module_request.response->modules[i];
         if (strcmp(initrd_module.path, file->path) == 0)
@@ -276,7 +276,7 @@ void _start()
 
         copy_mapping(boot_cr3, global_cr3, (uintptr_t)kernel_start_ptr, (uint64_t)((uintptr_t)kernel_end_ptr - (uintptr_t)kernel_start_ptr) >> 12);
 
-        for (int i = 0; i < mmap_request.response->entry_count; i++)
+        for (uint64_t i = 0; i < mmap_request.response->entry_count; i++)
         {
             struct limine_memmap_entry* entry = mmap_request.response->entries[i];
